added second valid isbn test

+++ src/test/java/com/mysy/isbn/validator/ValidateISBN.java
@@ -1,7 +1,7 @@
 package com.mysy.isbn.validator;
 
 public class ValidateISBN {
-    public boolean checkIsbn(int isbn) {
+    public boolean checkIsbn(long isbn) {
         if (isbn == 321125215) {
             return true;
         }
